+++ crates/air-lib/interpreter-interface/src/interpreter_outcome.rs
@@ -81,13 +81,7 @@ impl InterpreterOutcome {
         let error_message = try_as_string(record_values.pop().unwrap(), "error_message")?;
         let ret_code = try_as_i64(record_values.pop().unwrap(), "ret_code")?;
 
-        let outcome = Self {
-            ret_code,
-            error_message,
-            data,
-            next_peer_pks,
-            call_requests,
-        };
+        let outcome = Self::new(ret_code, error_message, data, next_peer_pks, call_requests);
 
         Ok(outcome)
     }