+++ avm/interface/CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.28.1] - 2022-12-26
+
+### Changed
+
+- Bump version of `avm-interpreter-interface` to 0.12.1:
+  it has removed a new `cid` field in the `InterpreterOutcome`.
+
+## [0.28.0] - 2022-12-21
+
+### Changed
+
+- Bump version of `avm-interpreter-interface` to 0.12.0:
+  it has a new `cid` field in the `InterpreterOutcome`.
+
 ## [0.27.0] - 2022-11-22
 
 ### Added